requirements.txt
Flask==2.2.2
Flask-WTF==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
telethon==1.24.0
requests==2.28.1
python-dotenv==0.20.0